Return Levels and Credible Intervals for a Poisson-GP Model

Compute return levels along with credible bounds for a minimal GEV model with Bayesian inference results.

Usage

# S3 method for GEVBayes0
RL(
  object,
  period = NULL,
  level = 0.7,
  credintType = c("HPD", "eqtail"),
  smooth = missing(period),
  ...
)

Arguments

object

An object with class GEVBayes0 representing the inference results for a GEV (block maxima) model.

period

A vector of periods for which the return levels will be computed.

level

The credible level

credintType

The type of credible interval wanted. See credInt.

smooth

Logical. If TRUE the bounds of the credible intervals are smoothed against the period.

...

Not used yet.

Value

An object with class "RL.GEVBayes" inheriting from "data.frame".

Period The Return Period. This is expressed in the same unit as the block duration that was given at the creation of object and which is stored as object$blockDuration. So if blockDuration is 2 (years) and Period is 100 (years), the Return Level is for \(50\) blocks.

Prob The probability of exceedance of the Return Level for the considered period. This is \(T / w^\star\) where \(T\) is the return period and \(w^\star\) is the block duration.

LevelThe credible level in formated form, e.g. "95%" for a provided level of 0.95.

Mode, Median, Mean While Median and Mean are the median and mean of the return levels \(\rho(T;\,\boldsymbol{\theta}^{[i]})\) corresponding to the MCMC iterates \(\boldsymbol{\theta}^{[i]}\) of the GEV parameter vector \(\boldsymbol{\theta}\), the mode is obtained by plugging the MAP of the GEV parameter into the return level \(\rho(T;\,\boldsymbol{\theta})\). The corresponding Return Level curve can be called "modal". If the MAP is not available in object, the corresponding column will contain NA.

L, U The Lower and Upper bounds of the credible interval.

Note that when \(m\) is a small integer \(>1\) and \(T = m w^\star\), the given probability is not the probability that the maximum over \(m\) blocks with duration \(w^\star\) exceeds the given level. This only holds when \(m\) is large.

References

Chapter 3 of

Coles S. (2001) An Introduction to Statistical Modeling of Extreme Values. Springer-Verlag.