Predictive Quantiles or Return Levels for a GEV Model, typically a Model for Block Maxima

Compute the predictive quantiles or return levels for a GEV Model. The quantiles are those for the maximum on a "new" period of time newDuration years.

Usage

# S3 method for GEVBayes0
predict(object, newDuration = 1.0, prob,
        type = "RL",
        approx = FALSE,
        trace = 0, ...)

Arguments

object

a GEVBayes0 object.

newDuration

The duration of the 'new' period for which the maximum is to be predicted. The newduration is expressed by using the block duration in object as unit. So newDuration = 10 means a duration of 10 * object$blockDuration.

prob

A vector of exceedance probabilities. The default value contains such as \(0.01\) and 0.001.

type

The type of prediction wanted. Remind that the predict method of the revdbayes package allows several types of prediction.

approx

Logical. For the default FALSE, each value of the tail quantile function is computed by zero-finding. For TRUE, the quantiles are computed by using a fine of pairs (argument, value) of the distribution function. This is likely to be faster than approx = FALSE when length(prob) is large.

trace

Integer level of verbosity.

...

Not used yet.

Value

A data frame with three columns

NewDuration

A vector indicating the duration of the predicted period.

Prob

The vector of exceedance probabilities.

Quant

The vector of quantile or return levels.

The dataframe is given the S3 class "predRL" and it receives several attributes such as the names of the factor columns.

Note

The duration newDuration is understood as given in years and not not in block duration. Thus it sould be kept constant when comparing Block Maxima models with different block durations, e.g. one and two years.

See also