Predictive Quantiles or Return Levels for a EV Model of type Poisson-GP

Prediction of an Extreme-Value model of type Poisson-GP using MCMC iterates.

Usage

# S3 method for poisGPBayes0
predict(object, newDuration = 1.0, prob,
        type = "RL",
        approx = FALSE,
        addLambda = !object$lambdaOut,
        trace = 0, ...)

Arguments

object

An poisGPBayes0 object, usually created by using the eponymous creator function poisGPBayes0.

newDuration

The duration of the 'new' period for which the maximum is to be predicted.

prob

A vector of exceedance probabilities. The default value contains probabilities such as 0.01 and 0.001.

type

The type of prediction wanted. Remind that the predict method of the revdbayes package uses this argument to allow several types of predictions: density "d", quantile "q", ...

approx

Logical. For the default FALSE, each value of the tail quantile function is computed by zero-finding. For TRUE, the quantiles are computed by using a fine of pairs (argument, value) of the distribution function. This is likely to be faster than approx = FALSE when length(prob) is large.

addLambda

Logical. If the description of object that 'lambda' is idependent of the GP parameters, the exact posterior of 'lambda' will by default be used to compute the predictive distribution. But if addLambda is then passed with ist value set to TRUE, a new colum corresponding to 'lambda' is simply added to the MCMC iterates and the computation is carried over by ignoring the specific independence property. In practice this seems to make little difference.

trace

Integer level of verbosity.

...

Not used yet.

Value

A data frame with the following columns

NewDuration

The duration of the "new" period on which the maximum is predicted.

Prob

A probability of exceedance.

Quantile

The return level corresponding to the probability.

The dataframe is given the S3 class "predRL" and it receives several attributes such as the names of the factor columns.

Details

The rate lambda can be omitted in the MCMC iterates of object, which will be reported via the logical flag lambdaOut set to TRUE among the elements of the object. Then, the parameter lambda is assumed to be a posteriori independent of the GP parameters. In this case, the predictive distribution can be computed by using the predictive distribution of the number of exceedances on the new period. This specific treatement can be by-passed by using addLamba = TRUE; in this case a column of MCMC iterates for lambda is added to the matrix of MCMC iterates for the GP parameters, see the Examples section of the poisGPBayes0 page.

See also